requirements.txt
appnope==0.1.2
backcall==0.2.0
decorator==5.1.0
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prettytable==2.2.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-irodsclient==1.0.0
six==1.16.0
snakeviz==2.1.0
tornado==6.1
traitlets==5.1.0
wcwidth==0.2.5
xmlrunner==1.7.7
